improve `task-parser` test.
