+++ task-parser/src/test/java/com/oppo/cloud/parser/utils/HDFSUtilTest.java
@@ -17,25 +17,36 @@
 package com.oppo.cloud.parser.utils;
 
 import com.oppo.cloud.common.domain.cluster.hadoop.NameNodeConf;
+import com.oppo.cloud.common.domain.job.LogPath;
 import com.oppo.cloud.parser.config.HadoopConfig;
 import com.oppo.cloud.parser.domain.reader.ReaderObject;
+import com.oppo.cloud.parser.service.reader.HDFSReader;
+import org.apache.hadoop.fs.FileStatus;
+import org.apache.hadoop.fs.LocatedFileStatus;
+import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.fs.RemoteIterator;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import javax.annotation.Resource;
+import java.io.IOException;
+import java.util.List;
 import java.util.Map;
 
-
 @SpringBootTest
-class HDFSUtilTest {
+class HDFSUtilTest extends ResourcePreparer {
 
     @Resource(name = HadoopConfig.NAME_NODE_MAP)
     Map<String, NameNodeConf> nameNodeMap;
 
-
-    void readLines()  {
-        try {
-            String path = "hdfs://logs-hdfs:8020/logs/application_1673850090992_23513";
+    @Test
+    void readLines() throws Exception {
+        String logDir = getTextLogDir();
+        RemoteIterator<LocatedFileStatus> files = getFileSystem().listFiles(new Path(logDir), true);
+        Assertions.assertTrue(files.hasNext());
+        while (files.hasNext()) {
+            String path = files.next().getPath().toString();
             NameNodeConf nameNode = HDFSUtil.getNameNode(nameNodeMap, path);
             ReaderObject readerObject = HDFSUtil.getReaderObject(nameNode, path);
             while (true) {
@@ -43,11 +54,8 @@ class HDFSUtilTest {
                 if (line == null) {
                     break;
                 }
-                System.out.println(line);
+                Assertions.assertTrue(!line.isEmpty());
             }
-        } catch (Exception e) {
-           e.printStackTrace();
         }
     }
-
 }
